+++ docs/docs/CHANGELOG.md
@@ -2,5 +2,6 @@
 
 ### Feat
 
+- add option for configuration file
 - first pass at the package
 - initialize the repository